rename AddPrayerConnected to AddPrayerScreenConnected
