+++ packages/apollos-ui-prayer/src/screens/index.js
@@ -1,2 +1,3 @@
+export AddPrayerScreenConnected from './AddPrayerScreenConnected';
 export ConfirmationDialogScreen from './ConfirmationDialogScreen';
 export PrayerDialogScreen from './PrayerDialogScreen';
